refactor: move References component hooks to hooks.ts file
