+++ src/components/references/hooks.ts
@@ -1,0 +1,14 @@
+import { create } from "zustand";
+
+export type State = {
+  references: string;
+};
+
+type Action = {
+  updateReferences: (description: State["references"]) => void;
+};
+
+export const useStore = create<State & Action>()((set) => ({
+  references: "",
+  updateReferences: (references) => set(() => ({ references: references })),
+}));
